pyproject.toml
[project]
name = "giscatalog"
version = "0.1.0"
description = "REST API application for the directory of Organizations, Buildings and Activities"
authors = [
    {name = "nadezhda",email = "nadezhda13popova@gmail.com"}
]

requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "poetry-core (>=2.3.2,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "aiohttp (>=3.13.5,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "geoalchemy2 (>=0.18.4,<0.19.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "sentry-sdk[fastapi] (>=2.57.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]
lint = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
    __pycache__
)/
"""

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
skip_gitignore = true
